moved dialog
